Code quality: exit cycle, util location, docs, decision cycle comment
1. Review /api/exit uses direct resolve instead of resolveAndCycle —
   no dangling cycle left for the decision loop to block on
2. Move retainUnchangedViewedFiles from types.ts to utils/diffFiles.ts
3. Document createDecisionCycle promise identity invariant
4. Update AGENTS.md: correct annotate match key format, add idle status
   lifecycle for code review, add version history endpoints to annotate
   server table
5. Add session lifetime and timeout facts to decisions.md
